requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
pytz
azure-functions
azure-ai-formrecognizer
azure-core
openai
azure-storage-blob
azure-cosmosdb-table
azure-identity
Template
rapidfuzz
pandas
PyPDF2
xlrd
openpyxl
numpy
pandas
fuzzywuzzy
